[Improvement] Pass to 8Bits sub code type, little endian (for x86 and arm performance improvement) (client,server,sha...

+++ server/master/EpollClientLoginMaster.h
@@ -80,13 +80,13 @@ private:
     void messageParsingLayer(const char * const message) const;
     //have message without reply
     void parseMessage(const quint8 &mainCodeType,const char *data,const int &size);
-    void parseFullMessage(const quint8 &mainCodeType,const quint16 &subCodeType,const char *data,const int &size);
+    void parseFullMessage(const quint8 &mainCodeType,const quint8 &subCodeType,const char *data,const int &size);
     //have query with reply
     void parseQuery(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const int &size);
-    void parseFullQuery(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
+    void parseFullQuery(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
     //send reply
     void parseReplyData(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const int &size);
-    void parseFullReplyData(const quint8 &mainCodeType,const quint16 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
+    void parseFullReplyData(const quint8 &mainCodeType,const quint8 &subCodeType,const quint8 &queryNumber,const char *data,const int &size);
 
     void parseInputBeforeLogin(const quint8 &mainCodeType,const quint8 &queryNumber,const char *data,const int &size);
     void disconnectClient();
